refactor: separate middleware and record metrics
